WIP: Improve the configuration API
